Passing first reader model test
