--- app/models/read.rb
@@ -1,8 +1,0 @@
-class Read < ApplicationRecord
-  validates :title, presence: true
-  validates :url, presence: true, uniqueness: true
-
-  def record_read(params)
-
-  end
-end
