+++ db/migrate/20170109201746_create_readers_table.rb
@@ -1,6 +1,6 @@
-class CreateReadsTable < ActiveRecord::Migration[5.0]
+class CreateReadersTable < ActiveRecord::Migration[5.0]
   def change
-    create_table :reads do |t|
+    create_table :readers do |t|
       t.string :title
       t.string :url
       t.integer :counter, default: 1
